#1380 move ByText, WithText, ByAttribute to subpackage `selector`
