+++ src/main/java/com/codeborne/selenide/selector/Xpath.java
@@ -1,0 +1,5 @@
+package com.codeborne.selenide.selector;
+
+class Xpath {
+  static final String NORMALIZE_SPACE_XPATH = "normalize-space(translate(string(.), '\t\n\r\u00a0', '    '))";
+}
